Moved clock_time from w/i Condition to utilities.
