+++ lib/concurrent/atomic/condition.rb
@@ -1,3 +1,5 @@
+require 'concurrent/utility/clock_time'
+
 module Concurrent
 
   # Condition is a better implementation of standard Ruby ConditionVariable. The
@@ -42,13 +44,13 @@ module Concurrent
     # @param [Object] timeout nil means no timeout
     # @return [Result]
     def wait(mutex, timeout = nil)
-      start_time = clock_time
+      start_time = Concurrent::clock_time
       @condition.wait(mutex, timeout)
 
       if timeout.nil?
         Result.new(nil)
       else
-        Result.new(start_time + timeout - clock_time)
+        Result.new(start_time + timeout - Concurrent::clock_time)
       end
     end
 
@@ -64,22 +66,6 @@ module Concurrent
     def broadcast
       @condition.broadcast
       true
-    end
-
-    private
-
-    if defined?(Process::CLOCK_MONOTONIC)
-      def clock_time
-        Process.clock_gettime Process::CLOCK_MONOTONIC
-      end
-    elsif RUBY_PLATFORM == 'java'
-      def clock_time
-        java.lang.System.nanoTime() / 1_000_000_000.0
-      end
-    else
-      def clock_time
-        Time.now.to_f
-      end
     end
   end
 end
