+++ lib/concurrent/utilities.rb
@@ -1,3 +1,4 @@
+require 'concurrent/utility/clock_time'
 require 'concurrent/utility/processor_count'
 require 'concurrent/utility/timeout'
 require 'concurrent/utility/timer'
